+++ app/Http/Controllers/Admin/Market/DiscountController.php
@@ -19,18 +19,21 @@ class DiscountController extends Controller
     //coupons
     public function coupon()
     {
+        $this->authorize('index', Coupon::class);
         $discounts = Coupon::all();
         return view('admin.market.discount.coupon', compact('discounts'));
     }
 
     public function couponCreate()
     {
+        $this->authorize('create', Coupon::class);
         $users = User::all();
         return view('admin.market.discount.couponCreate', compact('users'));
     }
 
     public function couponStore(CouponReqeust $request)
     {
+        $this->authorize('create', Coupon::class);
         $inputs = $request->all();
         $inputs['valid_from'] = date('Y-m-d', intval(substr($request->valid_from, 0, 10)));
         $inputs['valid_until'] = date('Y-m-d', intval(substr($request->valid_until, 0, 10)));
@@ -50,12 +53,14 @@ class DiscountController extends Controller
 
     public function couponEdit(Coupon $discount)
     {
+        $this->authorize('update', Coupon::class);
         $users = User::all();
         return view('admin.market.discount.couponEdit', compact('discount', 'users'));
     }
 
     public function couponUpdate(CouponReqeust $request, Coupon $discount)
     {
+        $this->authorize('update', Coupon::class);
         $inputs = $request->all();
         $inputs['valid_from'] = date('Y-m-d', intval(substr($request->valid_from, 0, 10)));
         $inputs['valid_until'] = date('Y-m-d', intval(substr($request->valid_until, 0, 10)));
@@ -74,6 +79,7 @@ class DiscountController extends Controller
 
     public function couponDestroy(Coupon $discount)
     {
+        $this->authorize('delete', Coupon::class);
         $discount->delete();
         return redirect()->route('admin.market.discount.coupon')->with("alertify-success", "کوپن موردنظر با موفقیت حذف شد");
     }
@@ -82,16 +88,19 @@ class DiscountController extends Controller
     //public discounts
     public function publicDiscount()
     {
+        $this->authorize('index', PublicDiscount::class);
         $discounts = PublicDiscount::all();
         return view('admin.market.discount.publicDiscount', compact('discounts'));
     }
 
     public function publicDiscountCreate()
     {
+        $this->authorize('create', PublicDiscount::class);
         return view('admin.market.discount.publicDiscountCreate');
     }
     public function publicDiscountStore(PublicDiscoutRequest $request)
     {
+        $this->authorize('create', PublicDiscount::class);
         $inputs = $request->all();
         $inputs['valid_from'] = date('Y-m-d', intval(substr($request->valid_from, 0, 10)));
         $inputs['valid_until'] = date('Y-m-d', intval(substr($request->valid_until, 0, 10)));
@@ -101,10 +110,12 @@ class DiscountController extends Controller
     }
     public function publicDiscountEdit(PublicDiscount $discount)
     {
+        $this->authorize('update', PublicDiscount::class);
         return view('admin.market.discount.publicDiscountEdit', compact('discount'));
     }
     public function publicDiscountUpdate(PublicDiscoutRequest $request, PublicDiscount $discount)
     {
+        $this->authorize('update', PublicDiscount::class);
         $inputs = $request->all();
         $inputs['valid_from'] = date('Y-m-d', intval(substr($request->valid_from, 0, 10)));
         $inputs['valid_until'] = date('Y-m-d', intval(substr($request->valid_until, 0, 10)));
@@ -114,6 +125,7 @@ class DiscountController extends Controller
     }
     public function publicDiscountDestroy(PublicDiscount $discount)
     {
+        $this->authorize('delete', PublicDiscount::class);
         $discount->delete();
         return redirect()->route('admin.market.discount.public')->with("alertify-success", "تخفیف موردنظر با موفقیت حذف شد");
     }
@@ -123,18 +135,21 @@ class DiscountController extends Controller
     //amazing sales
     public function amazingDiscount()
     {
+        $this->authorize('index', AmazingSale::class);
         $discounts = AmazingSale::all();
         return view('admin.market.discount.amazingDiscount', compact('discounts'));
     }
 
     public function amazingDiscountCreate()
     {
+        $this->authorize('create', AmazingSale::class);
         $products = Product::all();
         return view('admin.market.discount.amazingDiscountCreate', compact('products'));
     }
 
     public function amazingDiscountStore(AmazingSaleReqeust $request)
     {
+        $this->authorize('create', AmazingSale::class);
         $inputs = $request->all();
         $inputs['valid_from'] = date('Y-m-d', intval(substr($request->valid_from, 0, 10)));
         $inputs['valid_until'] = date('Y-m-d', intval(substr($request->valid_until, 0, 10)));
@@ -144,11 +159,13 @@ class DiscountController extends Controller
     }
     public function amazingDiscountEdit(AmazingSale $discount)
     {
+        $this->authorize('update', AmazingSale::class);
         $products = Product::all();
         return view('admin.market.discount.amazingDiscountEdit', compact('discount', 'products'));
     }
     public function amazingDiscountUpdate(AmazingSaleReqeust $request, AmazingSale $discount)
     {
+        $this->authorize('update', AmazingSale::class);
         $inputs = $request->all();
         $inputs['valid_from'] = date('Y-m-d', intval(substr($request->valid_from, 0, 10)));
         $inputs['valid_until'] = date('Y-m-d', intval(substr($request->valid_until, 0, 10)));
@@ -158,6 +175,7 @@ class DiscountController extends Controller
     }
     public function amazingDiscountDestroy(AmazingSale $discount)
     {
+        $this->authorize('delete', AmazingSale::class);
         $discount->delete();
         return redirect()->route('admin.market.discount.amazing')->with("alertify-success", "تخفیف شگفت انگیز موردنظر با موفقیت حذف شد");
     }
